Group Avatar: refactor and clean up
