+++ bp-rest.php
@@ -59,8 +59,9 @@ add_action( 'rest_api_init', function() {
 		$controller = new BP_REST_Members_Endpoint();
 		$controller->register_routes();
 
-		require_once( dirname( __FILE__ ) . '/includes/bp-attachments/classes/class-bp-rest-avatar-endpoint.php' );
-		$controller = new BP_REST_Attachments_Avatar_Endpoint();
+		require_once( dirname( __FILE__ ) . '/includes/bp-attachments/classes/trait-attachments.php' );
+		require_once( dirname( __FILE__ ) . '/includes/bp-attachments/classes/class-bp-rest-member-avatar-endpoint.php' );
+		$controller = new BP_REST_Attachments_Member_Avatar_Endpoint();
 		$controller->register_routes();
 	}
 
@@ -101,8 +102,9 @@ add_action( 'rest_api_init', function() {
 		$controller = new BP_REST_Group_Invites_Endpoint();
 		$controller->register_routes();
 
-		require_once( dirname( __FILE__ ) . '/includes/bp-attachments/classes/class-bp-rest-group-cover-endpoint.php' );
-		$controller = new BP_REST_Attachments_Group_Cover_Endpoint();
+		require_once( dirname( __FILE__ ) . '/includes/bp-attachments/classes/trait-attachments.php' );
+		require_once( dirname( __FILE__ ) . '/includes/bp-attachments/classes/class-bp-rest-group-avatar-endpoint.php' );
+		$controller = new BP_REST_Attachments_Group_Avatar_Endpoint();
 		$controller->register_routes();
 	}
